Create setupChromedriver method in utils
